Add some more

+++ parsers/c++-toml11.zsh
@@ -1,8 +1,28 @@
+# C++ compiler is slower than a blind crippled tortoise, so we need all these
+# stupid tricks if we don't want to wait 40 seconds...
 setup() {
-	c++ -I. -std=c++17 -O2 tests/check_toml_test.cpp -o check_toml_test
+	if [[ ! -e check_toml_test ]] || [[ tests/check_toml_test.cpp -nt check_toml_test ]]; then
+		c++ -I. -std=c++17 -O2 tests/check_toml_test.cpp -o check_toml_test
+	fi
 
-	cp ../../scripts/c++-toml11-perf.cpp .
-	c++ -I. -std=c++17 -O2 c++-toml11-perf.cpp -o perf
+	if [[ ! -e c++-toml11-perf.cpp ]] || \
+		[[ "$(sha256sum <<<$(< c++-toml11-perf.cpp))" != "$(sha256sum <<<$(< ../../scripts/c++-toml11-perf.cpp))" ]]; then
+		cp ../../scripts/c++-toml11-perf.cpp .
+	fi
+	if [[ ! -e perf ]] || [[ c++-toml11-perf.cpp -nt perf ]]; then
+		c++ -I. -std=c++17 -O2 c++-toml11-perf.cpp -o perf
+	fi
+
+	if [[ ! -e json.hpp ]]; then
+		curl -sOL https://raw.githubusercontent.com/nlohmann/json/develop/single_include/nlohmann/json.hpp
+	fi
+	if [[ ! -e c++-toml11-encoder.cpp ]] || \
+		[[ "$(sha256sum <<<$(< c++-toml11-encoder.cpp))" != "$(sha256sum <<<$(< ../../scripts/c++-toml11-encoder.cpp))" ]]; then
+		cp ../../scripts/c++-toml11-encoder.cpp .
+	fi
+	if [[ ! -e encoder ]] || [[ c++-toml11-encoder.cpp -nt encoder ]]; then
+		c++ -I. -std=c++17 -O2 c++-toml11-encoder.cpp -o encoder
+	fi
 }
 
 typeset -A info=(
